requirements.txt
dask==2023.2.0
pandas==1.5.3
pip-chill==1.0.1
plotly==5.13.0
pyarrow==11.0.0
python-levenshtein==0.20.9
tqdm==4.64.1